docs: Updates steps and repo in 3.2.1 so things work.
The main change here is from the previously linked (and rather bloated) internal repo devops-knowledge-share-ui.
For a lesson on hosting a static website first it did not work and additionally it was overly complex leading to confusion.
This branch also addressed some typos and poor assumptions that the previous steps made.
This branch also updates the S3 bucket permissions away from ACL rules (deprecated) and to Bucket Policies.